feat: Context API로 상태를 전역에서 관리하도록 구현
