+++ movie-project/pages/Main.jsx
@@ -1,24 +1,12 @@
+import { useContext } from "react";
 import MovieCard from "../component/MovieCard";
-import useFetch from "../hooks/useFetch";
+import { MovieContext } from "../context/MovieContext";
 
-const options = {
-  method: "GET",
-  headers: {
-    accept: "application/json",
-    Authorization:
-      "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJlODI2ZjVkOTZmMDY3ZjE2MmFkYWViZGQzZGE3ZGJmNyIsIm5iZiI6MTc1Njk1ODU4My4wODQsInN1YiI6IjY4YjkwZjc3MTU4ZTU2NjkyYTliYjI1NSIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.wYBV7FzivAFbGNs2coFwwVz0M0fJhx3J4HatiSC345I",
-  },
-};
 
 export default function Main() {
-  // const [movie,setMovie] = useState(data)
-  const [data, loading, error] = useFetch(
-    "https://api.themoviedb.org/3/movie/popular?language=en-US&page=1",
-    options
-  );
+  const {data,loading,error} = useContext(MovieContext)
 
   const movieForAll = data?.results.filter(el=>el.adult === false)
-  console.log(movieForAll)
 
   if(error)return <div className="text-center font-bold text-3xl pt-10">Error : 페이지를 찾 을 수 없습니다.</div>;
